+++ packages/at_client/test/encryption_service_test.dart
@@ -7,7 +7,6 @@ import 'package:at_client/src/encryption_service/self_key_encryption.dart';
 import 'package:at_client/src/encryption_service/shared_key_encryption.dart';
 import 'package:at_client/src/transformer/request_transformer/put_request_transformer.dart';
 import 'package:at_commons/at_builders.dart';
-import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
@@ -29,8 +28,6 @@ class MockAtClient extends Mock implements AtClient {
 class FakeLocalLookUpVerbBuilder extends Fake implements LLookupVerbBuilder {}
 
 class FakeAtSigningInput extends Fake implements AtSigningInput {}
-
-class MockCommitLogKeystore extends Mock implements CommitLogKeyStore {}
 
 void main() {
   LocalSecondary mockLocalSecondary = MockLocalSecondary();
@@ -280,7 +277,6 @@ void main() {
   group('A group of test related shared key encryption', () {
     String storageDir = '${Directory.current.path}/test/hive';
     SharedKeyEncryption sharedKeyEncryption;
-    AtCommitLog? atCommitLog;
 
     var encryptionPrivateKey =
         'MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCJ+42pSjfJ5DSe7jCh7RWSH9MlTiG3PPrYdEGHDoSNhtPflSiL6BhpEpodNMLSaYpWsoV9ad2vRgXqgN0qM3LufflkgAigpAU8ukzCUWs+7sHQZBPfiz/clO1GuajF7iV0CaOe7tJA7Hyod9StALDi/56kwhGqLi84rimLeNKxv7qCfOJowePJCrs5++KXQozQXWYoeeBrj1HGEWMCbuoaw3ihvwJPBo76GkR7seYcq9AlrFivf2HcvDaPx0fSMXejTS5aL0Kogz7hBrNNGi1WyjD51hZEvMssh1OYy6TKtYsj3veMO1zd9fBZT+9yXXNCZ3cXXnZqOHAetWEPsjoNAgMBAAECggEAQ7Bp4ECOebY/si+7H9SEnniKRmS72X5KuGDfvHd8w0j/K1Gq4Gdtgi4j+Gvnnv0zZjCRl+KVY+SABnhNBuTSXvjhnVHJ6bRM9WuXOERkziymW6qcrS9MltNgSy/NAbxAF1qbL96MuljJFoQiivQp0lH/62dg7xFVDQMzUj5lbdid0CIV9r9Jp6SxXCG8OTjhU8RupXyqWiIgg3xVr/Y2ayYnO2O8cnm9V8C0zTvX5290IijTfHXkbarGg0vY7sg2Krs1uk1sQ+70yHJSppTOfa+cfYLuoj4k+emqJWWn7TZKkTlfnc6ON566XN5eiH+t2AmKttW9MxWQhaWv50I7KQKBgQDH4yh3u8f+Pbj3L3SJMN0aQ/ZHMApJ96C+iAJRgYp0aYuVpHscQPRxS7p3JwT758cKdrUTHSji/t7PrtsLDlp9rLw8WHEng7rfA/hlL9Ghoks/3gxaX90+dCXMo1odKbCxTEEBGxUC0aFIOOQGJ8Ot47fNE15IuaoPDO/On1X36wKBgQCwt6H02nLxKY0ILdu8HRVGXcqOOcBQiMIrjjhmt7llAfB5cJKuwq29apyJdRglqWSHXuK3B42reiF3I0fm3+QfihqFUfchU+2N8LcCciNR1BM0l1t/Xkw/FW18lTld0WoTAI/EOE+VEOzzdO542f2m7EBjQZy9hVg4XtlKi1pP5wKBgCXrqVS1siZAbWOvhAs20utVs1Yj/f+0U7FxugbebXbSQyHbd2OPyw/nTvOl2mMzwGXyyT1cDdKqiXia8oExcudeqsNEAAuACSaf6TLBFKL2WBJAvNU0VJOxky40WzcnHpc0ISzlh2HmhRNff5rPVmcZyVfFceCYIHQEf0YSokuLAoGAFqnmXnWpqh4vFS50cOK1+MlMkgL8FBgF9voNZ7cGUtr10U1LspgLGjDTFJns19+qoeXcY6bXV3eZVSM0NHrgUd8vWYvSivatj7egcPLcbsEpGWST+njIhIql+QVWTx7tYLSAu6SRKEf8a5jCgMNMUZ0ZAOHITVINp2UarwHCOl8CgYBenz32mGQapDgw7fyw3aWe5e4i2rC3jHOxJOHSHTAcLzBZ7JrKOIoNtiHU9XWIjRB0kng4a0rDffywxM4YHI+ZMXgvYzHE1Ob2ei3Q/Q52bxkLEwEGrqwXl1kdmCb69imp1T92JBZ3ls2dX7+uJtJiy5KlZilGN61AwMgOxyg7Mg==';
@@ -312,21 +308,11 @@ void main() {
         print('DeleteVerbBuilder: ${builder.buildCommand()}');
         return 'data:10';
       });
-
-      atCommitLog = await AtCommitLogManagerImpl.getInstance().getCommitLog(
-          '@alice',
-          commitLogPath: storageDir,
-          enableCommitId: false);
     });
 
     test('test to verify legacy encryption when shared key is available',
         () async {
       sharedKeyEncryption = SharedKeyEncryption(mockAtClient);
-      await atCommitLog?.commitLogKeyStore.add(
-          // Adding commit id to mock commit entry is synced from server
-          CommitEntry('@bob:shared_key@alice', CommitOp.UPDATE, DateTime.now())
-            ..commitId = 0);
-      sharedKeyEncryption.atCommitLog = atCommitLog;
       var atKey = (AtKey.shared('phone', namespace: 'wavi', sharedBy: '@alice')
             ..sharedWith('@bob'))
           .build();
@@ -365,11 +351,6 @@ void main() {
 
     test('test to verify encryption when shared key is available', () async {
       sharedKeyEncryption = SharedKeyEncryption(mockAtClient);
-      await atCommitLog?.commitLogKeyStore.add(
-          // Adding commit id to mock commit entry is synced from server
-          CommitEntry('@bob:shared_key@alice', CommitOp.UPDATE, DateTime.now())
-            ..commitId = 0);
-      sharedKeyEncryption.atCommitLog = atCommitLog;
       var atKey = (AtKey.shared('phone', namespace: 'wavi', sharedBy: '@alice')
             ..sharedWith('@bob'))
           .build();
@@ -403,11 +384,6 @@ void main() {
     test('test to verify legacy encryption when a new shared key is generated',
         () async {
       sharedKeyEncryption = SharedKeyEncryption(mockAtClient);
-      // Adding commit id to mock that commit entry is synced from server
-      await atCommitLog?.commitLogKeyStore.add(
-          CommitEntry('@bob:shared_key@alice', CommitOp.UPDATE, DateTime.now())
-            ..commitId = 0);
-      sharedKeyEncryption.atCommitLog = atCommitLog;
 
       when(() => mockLocalSecondary
               .executeVerb(any(that: LLookupEncryptedSharedKeyMatcher())))
@@ -455,11 +431,6 @@ void main() {
     test('test to verify encryption when a new shared key is generated',
         () async {
       sharedKeyEncryption = SharedKeyEncryption(mockAtClient);
-      // Adding commit id to mock that commit entry is synced from server
-      await atCommitLog?.commitLogKeyStore.add(
-          CommitEntry('@bob:shared_key@alice', CommitOp.UPDATE, DateTime.now())
-            ..commitId = 0);
-      sharedKeyEncryption.atCommitLog = atCommitLog;
 
       when(() => mockLocalSecondary
               .executeVerb(any(that: LLookupEncryptedSharedKeyMatcher())))
@@ -504,11 +475,7 @@ void main() {
     test(
         'test to verify exception is thrown when update command fails to store shared_key to remote secondary',
         () async {
-      await atCommitLog!.commitLogKeyStore.add(CommitEntry(
-          '@bob:shared_key@alice', CommitOp.UPDATE, DateTime.now()));
-
       sharedKeyEncryption = SharedKeyEncryption(mockAtClient);
-      sharedKeyEncryption.atCommitLog = atCommitLog;
 
       when(() => mockLocalSecondary
               .executeVerb(any(that: LLookupEncryptedSharedKeyMatcher())))
@@ -544,7 +511,6 @@ void main() {
     });
 
     tearDown(() async {
-      await AtCommitLogManagerImpl.getInstance().close();
       var isExists = await Directory(storageDir).exists();
       if (isExists) {
         Directory(storageDir).deleteSync(recursive: true);